requirements.txt
albumentations
demucs
librosa
numpy
opencv-python
matplotlib
pandas
pytorch-lightning
pyyaml
rich
scikit-learn
scipy
torch
torchaudio
torchvision
tqdm
wandb
